add partonTop index, rename files to follow CamelCase convention
