+++ NanoAODProduction/python/customise_cff.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
 
 from TZWi.NanoAODProduction.customise_dropNanoAODTables_cff import *
-from TZWi.NanoAODProduction.customise_particletop_cff import *
+from TZWi.NanoAODProduction.customise_particleTop_cff import *
+from TZWi.NanoAODProduction.customise_partonTop_cff import *
